+++ src/zocalo/cli/dlq_purge.py
@@ -19,6 +19,7 @@ from functools import partial
 import workflows
 
 import zocalo.configuration
+from zocalo.util.rabbitmq import RabbitMQAPI
 
 
 def run() -> None:
@@ -64,11 +65,9 @@ def run() -> None:
     def receive_dlq_message(header: dict, message: dict, rabbitmq=False) -> None:
         idlequeue.put_nowait("start")
         if rabbitmq:
-            msg_time = (
-                int(datetime.timestamp(header["headers"]["x-death"][0]["time"])) * 1000
-            )
-            header["headers"]["x-death"][0]["time"] = datetime.timestamp(
-                header["headers"]["x-death"][0]["time"]
+            msg_time = int(datetime.timestamp(header["x-death"][0]["time"])) * 1000
+            header["x-death"][0]["time"] = datetime.timestamp(
+                header["x-death"][0]["time"]
             )
         else:
             msg_time = int(header["timestamp"])
@@ -109,16 +108,16 @@ def run() -> None:
         print(
             f"Message {header['message-id']} ({time.strftime('%Y-%m-%d %H:%M:%S', timestamp)}) exported:\n {os.path.join(filepath, filename)}"
         )
-        if rabbitmq:
-            # subscription_id does nothing for RabbitMQ but it is currently required by workflows
-            transport.ack(header, subscription_id=header["message-id"])
-        else:
-            transport.ack(header)
+        transport.ack(header)
         idlequeue.put_nowait("done")
 
     transport.connect()
     if not queues:
-        queues = [dlqprefix + ".>"]
+        if args.transport == "StompTransport":
+            queues = [dlqprefix + ".>"]
+        elif args.transport == "PikaTransport":
+            rmq = RabbitMQAPI.from_zocalo_configuration(zc)
+            queues = [q.name for q in rmq.queues() if q.name.startswith("dlq.")]
     for queue_ in queues:
         print("Looking for DLQ messages in " + queue_)
         transport.subscribe(